requirements.txt
colorama==0.4.6
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0.post0
Jinja2==3.1.2
login==0.0.6
SQLAlchemy==1.4.42
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
requests==2.22.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
